# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vena - Gestão Inteligente de Compras e Obras
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchases/requests:
    get:
      operationId: listPurchaseRequests
      tags: [purchases]
      summary: List purchase requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: project_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of purchase requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseRequest"
    post:
      operationId: createPurchaseRequest
      tags: [purchases]
      summary: Create purchase request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
  /purchases/requests/{id}:
    get:
      operationId: getPurchaseRequest
      tags: [purchases]
      summary: Get purchase request by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
    patch:
      operationId: updatePurchaseRequest
      tags: [purchases]
      summary: Update purchase request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"

  # QUOTES
  /purchases/quotes:
    get:
      operationId: listQuotes
      tags: [purchases]
      summary: List quotes
      parameters:
        - name: request_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [purchases]
      summary: Create quote from supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /purchases/quotes/{id}/approve:
    post:
      operationId: approveQuote
      tags: [purchases]
      summary: Approve quote and generate purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote approved and PO generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchases/quotes/compare:
    get:
      operationId: compareQuotes
      tags: [purchases]
      summary: Compare quotes for a purchase request
      parameters:
        - name: request_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteComparison"

  # PURCHASE ORDERS
  /purchases/orders:
    get:
      operationId: listPurchaseOrders
      tags: [purchases]
      summary: List purchase orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: supplier_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
  /purchases/orders/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [purchases]
      summary: Get purchase order by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchases/history:
    get:
      operationId: getPurchaseHistory
      tags: [purchases]
      summary: Get purchase history with price comparison
      parameters:
        - name: material
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Purchase history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseHistoryEntry"

  # FINANCIAL
...
components:
  schemas:
...
    PurchaseRequest:
      type: object
      required: [id, title, requestedBy, projectId, status, items, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        requestedBy:
          type: string
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, quoting, approved, ordered, delivered, cancelled]
        urgency:
          type: string
          enum: [low, normal, high, urgent]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseRequestItem"
        createdAt:
          type: string
    PurchaseRequestItem:
      type: object
      required: [id, materialName, quantity, unit]
      properties:
        id:
          type: integer
        materialName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        notes:
          type: ["string", "null"]
    PurchaseRequestInput:
      type: object
      required: [title, items]
      properties:
        title:
          type: string
        projectId:
          type: integer
        urgency:
          type: string
          enum: [low, normal, high, urgent]
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseRequestItemInput"
    PurchaseRequestItemInput:
      type: object
      required: [materialName, quantity, unit]
      properties:
        materialName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        notes:
          type: string
    PurchaseRequestUpdate:
      type: object
      properties:
        status:
          type: string
        urgency:
          type: string
        notes:
          type: string

    # QUOTES
    Quote:
      type: object
      required: [id, purchaseRequestId, supplierId, supplierName, totalAmount, deliveryDays, status, createdAt]
      properties:
        id:
          type: integer
        purchaseRequestId:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        totalAmount:
          type: number
        deliveryDays:
          type: integer
        freightCost:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        createdAt:
          type: string
    QuoteItem:
      type: object
      required: [id, materialName, quantity, unit, unitPrice, total]
      properties:
        id:
          type: integer
        materialName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        total:
          type: number
    QuoteInput:
      type: object
      required: [purchaseRequestId, supplierId, deliveryDays, items]
      properties:
        purchaseRequestId:
          type: integer
        supplierId:
          type: integer
        deliveryDays:
          type: integer
        freightCost:
          type: number
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItemInput"
    QuoteItemInput:
      type: object
      required: [materialName, quantity, unit, unitPrice]
      properties:
        materialName:
          type: string
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
    QuoteComparison:
      type: object
      required: [purchaseRequestId, quotes, recommendedQuoteId]
      properties:
        purchaseRequestId:
          type: integer
        quotes:
          type: array
          items:
            $ref: "#/components/schemas/Quote"
        recommendedQuoteId:
          type: ["integer", "null"]
        aiInsight:
          type: ["string", "null"]

    # PURCHASE ORDERS
    PurchaseOrder:
      type: object
      required: [id, quoteId, supplierId, supplierName, projectId, totalAmount, status, createdAt]
      properties:
        id:
          type: integer
        quoteId:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        expectedDelivery:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        createdAt:
          type: string
    PurchaseHistoryEntry:
      type: object
      required: [date, supplier, material, quantity, unit, unitPrice, total]
      properties:
        date:
          type: string
        supplier:
          type: string
        material:
          type: string
        quantity:
          type: number
        unit:
          type: string
        unitPrice:
          type: number
        total:
          type: number
        vsAvgPercent:
          type: ["number", "null"]

    # FINANCIAL
...